Adding usability script for agent blueprints
