+++ agents-for-bedrock/agent-blueprint-templates/cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "npx ts-node --prefer-ts-exts bin/03-agent-with-kb-and-guardrails/main.ts",
+  "app": "npx ts-node --prefer-ts-exts bin/main.ts",
   "watch": {
     "include": [
       "**"
